requirements.txt
numpy
opencv_python
Pillow
torch==2.11.2
torchvision